Move dbutil out of the scripts package
